Modify CharbonnierLoss, add ElasticLoss
In the original code, the CharbonnierLoss had no normalization, so the loss was 6 orders of magnitude higher than normal L1 loss under the same conditions. Added a normalization factor similar to L1 and L2, considering all the points (C*B*H*W) that makes the loss very similar in magnitude. The default epsilon of 1e-6 appears to be working well. 
Also adding the ElasticLoss as an alternative loss function. Default alpha is 0.5, but may want to keep it between 0.2 and 0.3 to make it less sensitive to outliers (L2 component)